* 1.38h, 25 January 2006       + Process/Filters/Gaussian Blur now works with non-integer radii thanks         to code contributed by Kai Uwe Barthel.       + When running on Java 1.6, ImageJ opens 8-bit images using half as much         memory.       + Added the setOption("DisableUndo", true) macro function. On Java 1.6,         with this option enabled, ImageJ can open and process an 8-bit image         using one third as much memory.       + Added the Plugins/Utilities/Reset command, which replaces Plugins/         Utilities/Unlock Image.       + Added eight links to the ImageJ website to the Help menu.       + The "B&C", "Threshold" and "Results" windows remember their screen         locations.       + If no images are open, images less than one half the width of the         screen open centered on the screen.       + Added a Duplicate command to the Edit menu of the "Results" window and         to the drop down menu that appears when you right click on the         "Results" window.       + Added the shift-x keyboard shortcut for the Image/Crop command.       + Edit/Paste displays text on the system clipboard in a text window if         the internal clipboard is empty and there is not an image on the system         clipboard (example).       + Image windows no longer move to the left edge of the screen after being         cropped or resized.       + The Plugins/Utilities/ImageJ Properties command displays more         information.       + Added a constructor to the ParticleAnalyser class that accepts         minCircularity and maxCircularity arguments.       + Fixed a bug that sometimes caused the scrollbar at the bottom of stack         windows to blink on machines running Windows.       + Fixed a bug that could cause negative calculated mean values with 8-bit         images larger than 8MB.       + Fixed a bug that could cause an exception when the ParticleAnalzer         class was called from a plugin.

+++ ij/process/ImageStatistics.java
@@ -81,7 +81,7 @@ public class ImageStatistics implements Measurements {
 		for (int i=minThreshold; i<=maxThreshold; i++) {
 			count = histogram[i];
 			pixelCount += count;
-			sum += i*count;
+			sum += (double)i*count;
 			value = i;
 			sum2 += (value*value)*count;
 			if (count>maxCount) {
